turkey-power-trading/requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
pyarrow==14.0.2
openpyxl==3.1.2
schedule==1.2.1
tqdm==4.66.1
